+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Add 
+- [#216] Annotation for exposed ports
+
+### Removed
+- [#216] Exposing services
 
 ## [v3.0.3] - 2024-12-12
 ### Added